scripts/requirements.txt
fastapi
uvicorn
kafka-python
minio
pandas
scikit-learn
streamlit
matplotlib
apache-airflow
great-expectations
sqlalchemy
psycopg2-binary
pgadmin4
grafana

